Tweak software installation methods
Download VS Code from direct link because
it gets such frequent updates.

Get the new Github Desktop beta with Electron.

Get the current Node version instead of LTS
in order to give Hyper and MSSQL a better shot.

Don't bother with the Notepad++ plugin manager
because it's on the roadmap to be baked in.